InputSystem and SoundSystem bug patches

+++ Minigin/InputSystem/Keyboard/Keyboard.cpp
@@ -3,7 +3,8 @@
 
 #include "SDL3/SDL_keyboard.h"
 
-dae::Keyboard::Keyboard() {
+dae::Keyboard::Keyboard()
+{
 	const bool* kbState = SDL_GetKeyboardState(nullptr);
 	memcpy(m_currentState.data(), kbState, SDL_SCANCODE_COUNT * sizeof(uint8_t));
 }
@@ -16,17 +17,17 @@ void dae::Keyboard::ProcessInput()
 
 bool dae::Keyboard::GetButtonHeld(uint32_t code) const
 {
-	return m_currentState[code];
+	return m_currentState[uint8_t(code)];
 }
 
 bool dae::Keyboard::GetButtonReleased(uint32_t code) const
 {
-	return !m_currentState[code] && m_previousState[code];
+	return !m_currentState[uint8_t(code)] && m_previousState[uint8_t(code)];
 }	
 
 bool dae::Keyboard::GetButtonPressed(uint32_t code) const
 {
-	return m_currentState[code] && !m_previousState[code];
+	return m_currentState[uint8_t(code)] && !m_previousState[uint8_t(code)];
 }
 
 float dae::Keyboard::GetAxis1D(uint32_t) const
